ref(node): Move `_dtrace` to `internal/dtrace`
